requirements.txt
apache-airflow==2.8.0
pandas==2.1.4
pyarrow==14.0.2
pyspark==3.5.0
minio==7.2.0
psycopg2-binary==2.9.9